+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 - Fixed Safari-specific layouting issues with a text inside a digram.
+- Version information is outputted to the browser's console.
 
 ## [0.0.14] - 2018-01-08